+++ server/src/admin/dto/album.dto.ts
@@ -8,7 +8,7 @@ import {
   Min,
   ValidateNested,
 } from 'class-validator';
-import { SongDto } from './SongDto';
+import { SongDto } from './song.dto';
 import { Type } from 'class-transformer';
 
 export class AlbumDto {
