+++ server/src/album/dto/main-banner-response.dto.ts
