+++ server/src/album/dto/side-bar-response.dto.ts
