+++ server/src/room/room.controller.ts
@@ -16,8 +16,8 @@ import { RandomNameUtil } from '@/common/randomname/random-name.util';
 import { Request } from 'express';
 import { Album } from '@/album/album.entity';
 import { Song } from '@/song/song.entity';
-import { AlbumResponseDto } from '@/album/albumResponse.dto';
-import { SongResponseDto } from '@/song/songResponse.dto';
+import { AlbumResponseDto } from '@/album/dto/album-response.dto';
+import { SongResponseDto } from '@/song/dto/song-response.dto';
 import { plainToInstance } from 'class-transformer';
 import { SongRepository } from '@/song/song.repository';
 import { ORDER } from '@/common/constants/repository.constant';
